Migrate CycleGraph tests
