+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/IrDumpTestGenerated.java
@@ -44,6 +44,12 @@ public class IrDumpTestGenerated extends AbstractIrDumpTest {
     }
 
     @Test
+    @TestMetadata("CycleGraph.kt")
+    public void testCycleGraph() {
+      runTest("compiler-tests/src/test/data/dump/ir/cycles/CycleGraph.kt");
+    }
+
+    @Test
     @TestMetadata("SelfCycle.kt")
     public void testSelfCycle() {
       runTest("compiler-tests/src/test/data/dump/ir/cycles/SelfCycle.kt");
